formatted

+++ app/controllers/home_controller.rb
@@ -1,19 +1,19 @@
+# frozen_string_literal: true
+
 class HomeController < ApplicationController
-  layout "home"
+  layout 'home'
 
   before_action :delete_expired_stories
 
   def delete_expired_stories
-    stories = ActiveStorage::Attachment.where(name: "story")
+    stories = ActiveStorage::Attachment.where(name: 'story')
     stories.each do |story|
-        if Date.parse(story.created_at.strftime("%Y-%m-%d")) < Date.today
-          story.destroy
-        end
+      story.destroy if Date.parse(story.created_at.strftime('%Y-%m-%d')) < Date.today
     end
   end
 
   def home
-    @posts = Post.all.order("created_at DESC").paginate(page: params[:page],per_page: 1)
+    @posts = Post.all.order('created_at DESC').paginate(page: params[:page], per_page: 1)
     Rails.logger.info @posts
     respond_to do |format|
       format.html
@@ -23,16 +23,18 @@ class HomeController < ApplicationController
   end
 
   def search
-    @users = User.where("user_name LIKE ? OR name LIKE ?", "%#{params[:q]}%", "%#{params[:q]}%")
-    @posts = Post.where("description LIKE ? OR created_at LIKE ?", "%#{params[:q]}%","%#{params[:q]}%")
+    @users = User.where('user_name LIKE ? OR name LIKE ?', "%#{params[:q]}%", "%#{params[:q]}%")
+    @posts = Post.where('description LIKE ? OR created_at LIKE ?', "%#{params[:q]}%", "%#{params[:q]}%")
   end
 
   def searching
     users = User.ransack(user_name_cont: params[:q]).result(distinct: true)
-    users = users.map { |user| { name: user.user_name, url: "/search?q=#{params[:q]}",icon: url_for(user.user_profile) } }
-    posts = Post.ransack(description_cont: params[:q]).result(distinct: true).map{ |post| {name: post.description,url: "/search?q=#{params[:q]}",icon: url_for(post.posts.first)} }
-    render json: {users: users[0,5], posts: posts }
+    users = users.map do |user|
+      { name: user.user_name, url: "/search?q=#{params[:q]}", icon: url_for(user.user_profile) }
+    end
+    posts = Post.ransack(description_cont: params[:q]).result(distinct: true).map do |post|
+      { name: post.description, url: "/search?q=#{params[:q]}", icon: url_for(post.posts.first) }
+    end
+    render json: { users: users[0, 5], posts: posts }
   end
-
-
 end
